--- src/extract/utl/extract-module-attributes.mjs
@@ -1,38 +1,0 @@
-/* eslint-disable security/detect-object-injection */
-
-/**
- * Given a module string returns in an object
- * - the module name
- * - the protocol (when encoded in the string)
- * - the mimeType (when encoded in the string)
- *
- * See https://nodejs.org/api/esm.html#esm_urls
- *
- * would've loved to use url.URL here, but that doesn't extract the mime type
- * (if there's a default node API that does this I'm all ears)
- *
- * @param {string} pString
- * @returns {any}
- */
-export default function extractModuleAttributes(pString) {
-  let lReturnValue = { module: pString };
-  const lModuleAttributes = pString.match(
-    // eslint-disable-next-line security/detect-unsafe-regex, unicorn/no-unsafe-regex
-    /^((node:|file:|data:)(([^,]+),)?)(.+)$/
-  );
-  const lProtocolPosition = 2;
-  const lMimeTypePosition = 4;
-  const lModulePosition = 5;
-
-  if (lModuleAttributes) {
-    lReturnValue.module = lModuleAttributes[lModulePosition];
-    if (lModuleAttributes[lProtocolPosition]) {
-      lReturnValue.protocol = lModuleAttributes[lProtocolPosition];
-    }
-    if (lModuleAttributes[lMimeTypePosition]) {
-      lReturnValue.mimeType = lModuleAttributes[lMimeTypePosition];
-    }
-  }
-
-  return lReturnValue;
-}
